move DataDownload resources to its own file
