+++ webapp/pages/Assessment/DataDownload/resources.js
@@ -1,0 +1,25 @@
+const resources = [
+  {
+    idx: 1,
+    name: 'Forest_extent_characteristics_and_changes',
+    labelKey: 'navigation.sectionHeaders.forestExtentCharacteristicsAndChanges',
+  },
+  {
+    idx: 2,
+    name: 'Forest_growing_stock_biomass_and_carbon',
+    labelKey: 'navigation.sectionHeaders.forestGrowingStockBiomassAndCarbon',
+  },
+  {
+    idx: 3,
+    name: 'Forest_designation_and_management',
+    labelKey: 'navigation.sectionHeaders.forestDesignationAndManagement',
+  },
+  {
+    idx: 4,
+    name: 'Forest_ownership_and_management_rights',
+    labelKey: 'navigation.sectionHeaders.forestOwnershipAndManagementRights',
+  },
+  { idx: 6, name: 'Permanent_forest_estate', labelKey: 'areaOfPermanentForestEstate.areaOfPermanentForestEstate' },
+]
+
+export default resources
